Forgot conversion on non-count version of query

+++ src/lib.rs
@@ -513,12 +513,12 @@ pub fn create_filter(input: TokenStream) -> TokenStream {
                     }
                 };
 
-                let mut limit = self.limit.unwrap_or(100);
+                let mut limit = self.limit.unwrap_or(100) as i64;
                 if limit < 0 {
                     limit = 100;
                 }
 
-                let mut page = self.page.unwrap_or(0);
+                let mut page = self.page.unwrap_or(0) as i64;
                 if page < 0 {
                     page = 0;
                 }
